Created own file for ByteVal, DatenElementGruppe, Konto, NULL and Segment.
