+++ lib/ByteVal.js
@@ -1,0 +1,5 @@
+function ByteVal(ddd) {
+  this.data = ddd
+}
+
+module.exports = ByteVal
