+++ lib/Classes.js
@@ -32,27 +32,11 @@ var util = require('util')
 var Parser = require('./Parser')
 var ParseError =  require('./errors').ParseError
 
-var Konto = function () {
-  var me_konto = this
-  me_konto.iban = ''
-  me_konto.konto_nr = ''
-  me_konto.unter_konto_merkm = null
-  me_konto.ctry_code = ''
-  me_konto.blz = ''
-  me_konto.kunden_id = ''
-  me_konto.kontoart = ''
-  me_konto.currency = ''
-  me_konto.kunde1_name = ''
-  me_konto.product_name = ''
-}
-
-var NULL = new function () {
-  this.id = 1234
-}()
-
-var ByteVal = function (ddd) {
-  this.data = ddd
-}
+var Konto = require('./Konto')
+var NULL = require('./NULL')
+var ByteVal = require('./ByteVal')
+var DatenElementGruppe = require('./DatenElementGruppe')
+var Segment = require('./Segment')
 
 var Helper = new function () {
   this.checkMsgsWithBelongToForId = function (msg, bez, id) {
@@ -242,204 +226,6 @@ var Helper = new function () {
     return str.replace(/[?+:]/g, '?$&')
   }
 }()
-
-var DatenElementGruppe = function () {
-  var me_deg = this
-  me_deg.next_el = 0
-  me_deg.data = new Array()
-  me_deg.desc = new Array()
-
-  me_deg.addDE = function (val) {
-    me_deg.data[me_deg.next_el] = val
-    me_deg.desc[me_deg.next_el] = 1
-    me_deg.next_el++
-  }
-
-  me_deg.addDEbin = function (val) {
-    me_deg.data[me_deg.next_el] = val
-    me_deg.desc[me_deg.next_el] = 3
-    me_deg.next_el++
-  }
-
-  me_deg.addDEG = function (grup) {
-    me_deg.data[me_deg.next_el] = grup
-    me_deg.desc[me_deg.next_el] = 2
-    me_deg.next_el++
-  }
-
-  me_deg.parse = function (parser) {
-    var start_pos
-    var first = false
-    while (!first || (parser.getCurrentChar() == ':' && parser.hasNext())) {
-      if (!first) first = true
-      else parser.nextPos()
-      start_pos = parser.getCurrentPos()
-      parser.setMarkerWithCurrentPos('start')
-      if (parser.getCurrentChar() == '@') {
-        // binary
-        parser.nextPos()
-        parser.setMarkerWithCurrentPos('start')
-        if (!parser.gotoNextValidChar('@')) throw new ParseError('Seg', 'Error binary!', start_pos)
-        var len = parseInt(parser.getTextFromMarkerToCurrentPos('start'), 10)
-        parser.nextPos()
-        parser.setMarkerWithCurrentPos('start')
-        parser.setCurrentPos(parser.getCurrentPos() + len)
-        if ("+:'".indexOf(parser.getCurrentChar()) == -1) throw new ParseError('Seg', 'Error binary, Wrong Length!' + len, start_pos)
-        me_deg.addDEbin(parser.getTextFromMarkerToCurrentPos('start'))
-        parser.nextPos()
-      } else if (parser.gotoNextValidCharButIgnoreWith("+:'", '?')) {
-        // Normales datenelement
-        me_deg.addDE(parser.getTextFromMarkerToCurrentPos('start'))
-        // Datengruppe können nicht bestandteil einer datengruppe sein
-      } else {
-        throw new ParseError('Seg', 'Unerwartetes ENDE!', start_pos)
-      }
-    }
-  }
-
-  me_deg.transformForSend = function () {
-    var result = ''
-    for (var i = 0; i != me_deg.data.length; i++) {
-      if (me_deg.data[i] != NULL) {
-        if (me_deg.desc[i] == 1) {
-          result += (i != 0 ? ':' : '') + me_deg.data[i] // DE
-        } else if (me_deg.desc[i] == 2) { // kommt nicht vor
-          result += (i != 0 ? ':' : '') + me_deg.data[i].transformForSend() // DEG
-        } else if (me_deg.desc[i] == 3) {
-          result += (i != 0 ? ':' : '') + '@' + me_deg.data[i].length + '@' + me_deg.data[i] // BIN DAT
-        }
-      } else {
-        // leer
-        result += (i != 0 ? ':' : '')
-      }
-    }
-    return result
-  }
-
-  me_deg.getEl = function (i) {
-    return me_deg.data[i - 1]
-  }
-}
-
-var Segment = function () {
-  var me_seg = this
-  me_seg.name = null
-  me_seg.nr = 0
-  me_seg.vers = 0
-  me_seg.bez = 0
-  me_seg.store = new DatenElementGruppe()
-
-  me_seg.init = function (n, nr, ve, be) {
-    me_seg.name = n
-    me_seg.nr = nr
-    me_seg.vers = ve
-    me_seg.bez = be
-  }
-
-  me_seg.transformForSend = function () {
-    var result = ''
-    result += me_seg.name // Nr. 1 Segmentkennung an ..6 M 1
-    result += ':' + me_seg.nr // Nr. 2 Segmentnummer num ..3 M 1 >=1
-    result += ':' + me_seg.vers // Nr. 3 Segmentversion GD num ..3 M 1
-    if (me_seg.bez !== 0) result += ':' + me_seg.bez
-    for (var i = 0; i != me_seg.store.data.length; i++) {
-      if (me_seg.store.data[i] != NULL) {
-        if (me_seg.store.desc[i] == 1) {
-          result += '+' + me_seg.store.data[i] // DE
-        } else if (me_seg.store.desc[i] == 2) {
-          result += '+' + me_seg.store.data[i].transformForSend() // DEG
-        } else if (me_seg.store.desc[i] == 3) {
-          result += '+@' + me_seg.store.data[i].length + '@' + me_seg.store.data[i] // BIN DAT
-        }
-      } else {
-        // leer
-        result += '+'
-      }
-    }
-    result += "'"
-    return result
-  }
-
-  me_seg.parse = function (parser) {
-    var start_pos = parser.getCurrentPos()
-    // 1. Segmentkopf
-    // Nr. 1 Segmentkennung an ..6 M 1
-    parser.setMarkerWithCurrentPos('start')
-    if (parser.gotoNextValidChar(':')) {
-      me_seg.name = parser.getTextFromMarkerToCurrentPos('start')
-    } else {
-      throw new ParseError('Seg', 'Segmentkennung Fehlt!', start_pos)
-    }
-
-    // Nr. 2 Segmentnummer num ..3 M 1 >=1
-    parser.nextPos()
-    start_pos = parser.getCurrentPos()
-    parser.setMarkerWithCurrentPos('start')
-    if (parser.gotoNextValidChar(':')) {
-      me_seg.nr = parser.getTextFromMarkerToCurrentPos('start')
-    } else {
-      throw new ParseError('Seg', 'Segmentnummer fehlt!', start_pos)
-    }
-
-    // Nr. 3 Segmentversion GD num ..3 M 1
-    parser.nextPos()
-    start_pos = parser.getCurrentPos()
-    parser.setMarkerWithCurrentPos('start')
-    if (parser.gotoNextValidChar(":+'")) {
-      me_seg.vers = parser.getTextFromMarkerToCurrentPos('start')
-    } else {
-      throw new ParseError('Seg', 'Segmentversion fehlt!', start_pos)
-    }
-
-    // Nr. 4 Bezugssegment GD num ..3 K 1 >=1
-    if (parser.getCurrentChar() == ':') {
-      parser.nextPos()
-      start_pos = parser.getCurrentPos()
-      parser.setMarkerWithCurrentPos('start')
-      if (parser.gotoNextValidChar('+')) {
-        me_seg.bez = parser.getTextFromMarkerToCurrentPos('start')
-      } else {
-        throw new ParseError('Seg', 'Unerwartetes ENDE!', start_pos)
-      }
-    }
-
-    // jetzt kommen datenlemente oder datenelementgruppen
-    while (parser.getCurrentChar() != "'" && parser.hasNext()) {
-      parser.nextPos()
-      start_pos = parser.getCurrentPos()
-      parser.setMarkerWithCurrentPos('start')
-      if (parser.getCurrentChar() == '@') {
-        // binary
-        parser.nextPos()
-        parser.setMarkerWithCurrentPos('start')
-        if (!parser.gotoNextValidChar('@')) throw new ParseError('Seg', 'Error binary!', start_pos)
-        var len = parseInt(parser.getTextFromMarkerToCurrentPos('start'), 10)
-        parser.nextPos()
-        parser.setMarkerWithCurrentPos('start')
-        parser.setCurrentPos(parser.getCurrentPos() + len)
-        if ("+:'".indexOf(parser.getCurrentChar()) == -1) throw new ParseError('Seg', 'Error binary, Wrong Length!' + len, start_pos)
-        me_seg.store.addDEbin(parser.getTextFromMarkerToCurrentPos('start'))
-      } else if (parser.gotoNextValidCharButIgnoreWith("+:'", '?')) {
-        if (parser.getCurrentChar() == '+' || parser.getCurrentChar() == "'") {
-          // Normales datenelement
-          me_seg.store.addDE(parser.getTextFromMarkerToCurrentPos('start'))
-        } else {
-          // Datengruppe
-          parser.setPosBackToMarker('start')
-          var neu_deg = new DatenElementGruppe()
-          neu_deg.parse(parser)
-          me_seg.store.addDEG(neu_deg)
-        }
-      } else {
-        throw new ParseError('Seg', 'Unerwartetes ENDE!', start_pos)
-      }
-    }
-  }
-
-  me_seg.getEl = function (nr) {
-    return me_seg.store.data[nr - 1]
-  }
-}
 
 var Nachricht = function (proto_version) {
   var me_msg = this
