+++ lib/Konto.js
@@ -1,0 +1,15 @@
+function Konto() {
+  var me_konto = this
+  me_konto.iban = ''
+  me_konto.konto_nr = ''
+  me_konto.unter_konto_merkm = null
+  me_konto.ctry_code = ''
+  me_konto.blz = ''
+  me_konto.kunden_id = ''
+  me_konto.kontoart = ''
+  me_konto.currency = ''
+  me_konto.kunde1_name = ''
+  me_konto.product_name = ''
+}
+
+module.exports = Konto;
