+++ lib/NULL.js
@@ -1,0 +1,3 @@
+module.exports = new function () {
+  this.id = 1234
+}()
